Incorporate CTikz for immediate graph generation
I found a few bugs in my implementation when I tried to plot the graphs,
such as an off by one, which lead to recording of 0 bytes accessed
cachelines and forgetting about 325 cachelines that were fully utilized.
Further, this commit contains additional logic to track instructions,
and the basic toolset to produce different graph outputs.