Debugging and fixing "ghost misses"
Ghost misses stem from address aliasing in lower level inclusive caches
leading to the eviction of the code from one to thje other instruction.
(E.g. reading sequence from addresses with the same index as the current
code block, as instruction is in l1, but data is read from memory, hence
goes through llc and l2, thus LRU counters are worse for the code block
in l2, that at some point it might be replaced by data. Thus we can have
a miss in the next instruction.) We now account for those misses as well
and count them as "aliasing miss"