Improved naming and efficiency for reruns
If we hit a timelimit on the server, we might want to continue where we
left off. So instead of recalculating everything, we return early when
we notice that the output files already exist. The check if they are
complete has to be done manually though, as we do not yet have an
indicator of completeness.