Debugging LRU replacement never updating counters
Had to do with wrongly invaliidating and setting counters to 0 for all
entries. An edge case still exists, when a whole set consists of only
blocks that are having the same tag (which in theory should never happen
as we have more than one 64 CB but we need to keep this in mind if we
have one or 0 fully sized ways). Either we have to invalidate with an
increasing counter (rendering it more difficult to implement a new
replacement policy), or let the replacement policy choose the default
value for invalid entries and setting the correct entry on insertion
(which is usually already done)